refactor(mf): pulpit - scss into css
