--- micro-frontends/src/mf-pulpit/src/ui/modules/pulpit-item/pulpit-item.module.scss
@@ -1,16 +1,0 @@
-.item {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	flex-direction: column;
-	cursor: pointer;
-	@include font('medium', 11, 13);
-}
-
-.icon {
-	margin-bottom: 6px;
-}
-
-.name {
-	color: var(--text-color-2);
-}
