+++ micro-frontends/src/mf-pulpit/src/ui/modules/pulpit/pulpit.module.css
@@ -1,10 +1,10 @@
 .pulpit {
 	display: grid;
-	grid-column-gap: 5px;
-	grid-row-gap: 5px;
+	grid-column-gap: var(--ds-spacing-sm);
+	grid-row-gap: var(--ds-spacing-sm);
 	grid-template-columns: repeat(auto-fill, 75px);
 	grid-template-rows: repeat(auto-fill, 65px);
 	grid-auto-flow: column;
 	width: 100%;
-    height: 100%;
+	height: 100%;
 }
