--- micro-frontends/src/mf-pulpit/src/ui/modules/wrapper/wrapper.module.scss
@@ -1,6 +1,0 @@
-.wrapper {
-  width: 100vw;
-  height: calc(100vh - 28px);
-  background-color: var(--background-color-pulpit);
-  padding: 1px;
-}
